Update config.lua
Changed order so the talent specs are in the same order as they are in game.

Changed `requireTree' line to be on a consistent line and made sure it was on each bar for each spec, this should make it easier to split it up to specs instead of classes it that happens.

Combined some CD onto bars for shadow. Honestly not sure if that makes it better or not, it is less bars overall though.