Update config.lua
Complete paladin config for all specs.
I don't have a paladin but had a friend test it, he thought it seemed good so I've provided it here. Some tweaks might be needed but feedback is required.

Note, this will error if you use it with v1.9.11 due to `requiresArtifactTalent` not working until the next release.