Warlock config update.
Made formatting accurate.
Added/changed required level for a few spells as I made a new warlock and had three bars available, I had to work out what they are not knowing the icons. 
Changed spec orders to match in game order, not technically needed but it makes it easier to find what you want (for me at least) and led to the above change.