Hunter config update 1.
Changed formatting for consistency.
Added colours to many, if not all, bars.
Added Beast Master and Survival configs.
Built upon the existing Marksmanship config, adding a few things and moving the bars position. Trying to have the highest priority things at the top but this spec is tricky to make work for this spec.