Do not merge!
I've put this here so you can see what I commented out to get rid of errors in Bugsack, I also deleted the references to LibArtifactData from the `toc` file. After doing that it doesn't matter if the folder is there or not, but you probably know that.

While this gets rid of the errors I can't get it to display the config. I turned debugging on and got a screen full of info I didn't know how to interpret, then I thought I'd just make the config smaller with the one which really matters to me, feral. After doing that I noticed the debug info was much shorter, then I clicked.
I dropped it down to one simple debuff, it still wouldn't load. :-(
I did work out what the debug message was telling me though `| spellId | debuffId | buffId | tick time | cooldown |`.

With the one spell in the config if I type `/ehz status` the result is this:
```EventHorizon | DEBUG |  | Adding | 164812 | 164812 | nil | 2 |nil
EventHorizon | DEBUG |  |  config:1
EventHorizon | Installed plugins:
EventHorizon |  redshift
EventHorizon |  lines
EventHorizon | Visible bars:```

So it's recognising the spell in the config but it isn't making it to the draw section.