Stackfont and Movement Handle.
I've changed the "EventHorizonHandle" to the top left of the window, I did this on my local copy as I've been using WeakAuras2 as an overlay for Pandemic indicators.
Moving the default handle means I can share a WA config which will overlay in the correct place even if the user has changed the width of the window. Which is why I changed my local one as getting the icons in the same place was more frustrating than it should have been.

I've also set a default Stack Font, outline, colour and offset.
I did this as for some specs the stacks matter and it wasn't clear enough, I had to change the colour to one which clashed with most of the icons so it was visible.

While I haven't test every spec which has a stack count, which is important to see, I think this change is worthwhile as the default wasn't clearly visible on all specs anyway and was terrible on the specs I played where it did matter.
Many hours were put into trying the available fonts and then checking a few specs.