Priest config file for 8.0.1 patch.
Updated the bars to account for the talent position changes, additions or removals.
Added colour to most of the spells for all specs, the idea being that you'll get used to it and be able to tell what is coming off CD without specifically looking at it. I've tried to make the bar colour similar to the icon's but as Disc and Holy have so many with a similar colour I had to choose something.

Changed the bar orders a bit to fit a common theme.
Offensive spells first, even for healers as healer DPS seems to be something Blizzard want us to do in between healing. This is also what Event Horizon does best.
If spells have a CD the order is shortest first then working down, this works for healing spell CD too, they start after the DPS bars, which for priest is pretty quickly.

While Renew and Atonement are tracked via mouseover they don't have their own bars, Renew is on the _Prayed of Mending_ bar and Atonement is on the _Power Word: Radiance_ bar.
I've made this change to keep the bar count down and also because I'd be astounded if anyone uses EH this way, you'll have those buffs on your party/raid frames.
We can't track the number of players with Atonement, not sure if we'll even be able to do that.
I suspect most of you have a WA for it already, or a specific healing mod for it. ;-)