Rogue config for 8.0.1 update two. (#98)
* Rogue config for 8.0.1 update two.

Added Blade Flurry after Grinder made me aware of the changes in Discord.
Changed some colours around. The icons for this spec are all so similar so the colours don't match the icons as well. :-(
Changed the CD under RtB/S&D to be _Between the Eyes_ as you'll probably use it more in M+ than caring if _Sprint_ is on CD or not.
_Blade Rush_ buff is no tracked, however I couldn't get it to track with Alacrity. I could have added another bar but figured it wasn't worth it to track getting **5** energy every second for five seconds.

Changed default past and future time to 1 and 9 seconds, 12 feels a bit long for rogue and we can change this by class now.

* Rogue config 8.0.1 update three.

Changed future time to 11.
Changed alpha channel for _Between the Eyes_ and RtB/S&D to make it clearer.
Added some missing talents for Subtlety.