Druid config for 8.1 patch. (#112)
Updated the feral section to work with the 8.1 changes.
I haven't checked the other configs but they didn't have the same level of changes, I'll upload a revised one with all specs tested in the next day or two. At least with this here feral druids can grab an updated config, manually, before the next release.